upload working, not setting price only

+++ Application/src/main/java/com/kenzie/appserver/controller/model/ArtworkCreateRequest.java
@@ -36,7 +36,7 @@ public class ArtworkCreateRequest {
 
     @Min(0)
     @JsonProperty("price")
-    private Double price;
+    private int price;
 
     public String getArtistName() {
         return artistName;
@@ -86,11 +86,11 @@ public class ArtworkCreateRequest {
         this.forSale = forSale;
     }
 
-    public Double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 }
